[Refactor] Report 패키징 추가
