+++ keewe-api/src/main/java/ccc/keeweapi/service/report/ReportApiService.java
@@ -1,8 +1,8 @@
-package ccc.keeweapi.service.insight;
+package ccc.keeweapi.service.report;
 
 import ccc.keeweapi.component.ReportAssembler;
 import ccc.keeweapi.dto.insight.ReportRequest;
-import ccc.keewedomain.service.insight.ReportDomainService;
+import ccc.keewedomain.service.report.ReportDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,5 +15,4 @@ public class ReportApiService {
     public void reportInsight(ReportRequest request) {
         reportDomainService.save(reportAssembler.toReportCreateDto(request.getInsightId(), request.getReportType(), request.getReason()));
     }
-
 }
