+++ keewe-domain/src/main/java/ccc/keewedomain/persistence/repository/report/ReportRepository.java
@@ -1,4 +1,4 @@
-package ccc.keewedomain.persistence.repository.insight;
+package ccc.keewedomain.persistence.repository.report;
 
 import ccc.keewedomain.persistence.domain.insight.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
